+++ NexCheckbox.cpp
@@ -5,6 +5,8 @@
  *
  * @author  huang xiaoming (email:<xiaoming.huang@itead.cc>)
  * @date    2016/9/13
+ * @author Jyrki Berg 2/17/2019 (https://github.com/jyberg)
+ * 
  * @copyright 
  * Copyright (C) 2014-2015 ITEAD Intelligent Systems Co., Ltd. \n
  * This program is free software; you can redistribute it and/or
@@ -14,15 +16,15 @@
  */
 #include "NexCheckbox.h"
 
-NexCheckbox::NexCheckbox(uint8_t pid, uint8_t cid, const char *name)
-    :NexTouch(pid, cid, name)
+NexCheckbox::NexCheckbox(uint8_t pid, uint8_t cid, const char *name, const NexObject* page)
+    :NexTouch(pid, cid, name, page)
 {
 }
 
 uint32_t NexCheckbox::getValue(uint32_t *number)
 {
     String cmd = String("get ");
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".val";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -34,7 +36,7 @@ bool NexCheckbox::setValue(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".val=";
     cmd += buf;
 
@@ -46,7 +48,7 @@ uint32_t NexCheckbox::Get_background_color_bco(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".bco";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -58,14 +60,14 @@ bool NexCheckbox::Set_background_color_bco(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".bco=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd="";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -74,7 +76,7 @@ uint32_t NexCheckbox::Get_font_color_pco(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".pco";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -86,14 +88,14 @@ bool NexCheckbox::Set_font_color_pco(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".pco=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }