+++ NexGauge.cpp
@@ -5,6 +5,8 @@
  *
  * @author  Wu Pengfei (email:<pengfei.wu@itead.cc>)
  * @date    2015/8/13
+ * @author Jyrki Berg 2/17/2019 (https://github.com/jyberg)
+ * 
  * @copyright 
  * Copyright (C) 2014-2015 ITEAD Intelligent Systems Co., Ltd. \n
  * This program is free software; you can redistribute it and/or
@@ -15,15 +17,15 @@
 
 #include "NexGauge.h"
 
-NexGauge::NexGauge(uint8_t pid, uint8_t cid, const char *name)
-    :NexObject(pid, cid, name)
+NexGauge::NexGauge(uint8_t pid, uint8_t cid, const char *name, const NexObject* page)
+    :NexObject(pid, cid, name, page)
 {
 }
 
 bool NexGauge::getValue(uint32_t *number)
 {
     String cmd = String("get ");
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".val";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -35,7 +37,7 @@ bool NexGauge::setValue(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".val=";
     cmd += buf;
 
@@ -47,7 +49,7 @@ uint32_t NexGauge::Get_background_color_bco(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".bco";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -59,14 +61,14 @@ bool NexGauge::Set_background_color_bco(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".bco=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd="";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -75,7 +77,7 @@ uint32_t NexGauge::Get_font_color_pco(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".pco";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -87,14 +89,14 @@ bool NexGauge::Set_font_color_pco(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".pco=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -103,7 +105,7 @@ uint32_t NexGauge::Get_pointer_thickness_wid(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".wid";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -115,14 +117,14 @@ bool NexGauge::Set_pointer_thickness_wid(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".wid=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -131,7 +133,7 @@ uint32_t NexGauge::Get_background_cropi_picc(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".picc";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -143,14 +145,14 @@ bool NexGauge::Set_background_crop_picc(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".picc=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }