+++ NexHotspot.cpp
@@ -5,6 +5,8 @@
  *
  * @author  Wu Pengfei (email:<pengfei.wu@itead.cc>)
  * @date    2015/8/13
+ * @author Jyrki Berg 2/17/2019 (https://github.com/jyberg)
+ * 
  * @copyright 
  * Copyright (C) 2014-2015 ITEAD Intelligent Systems Co., Ltd. \n
  * This program is free software; you can redistribute it and/or
@@ -15,8 +17,8 @@
  
 #include "NexHotspot.h"
 
-NexHotspot::NexHotspot(uint8_t pid, uint8_t cid, const char *name)
-    :NexTouch(pid, cid, name)
+NexHotspot::NexHotspot(uint8_t pid, uint8_t cid, const char *name, const NexObject* page)
+    :NexTouch(pid, cid, name, page)
 {
 }