+++ NexNumber.cpp
@@ -5,6 +5,8 @@
  *
  * @author  huang xianming (email:<xianming.huang@itead.cc>)
  * @date    2015/8/13
+ * @author Jyrki Berg 2/17/2019 (https://github.com/jyberg)
+ * 
  * @copyright 
  * Copyright (C) 2014-2015 ITEAD Intelligent Systems Co., Ltd. \n
  * This program is free software; you can redistribute it and/or
@@ -14,15 +16,15 @@
  */
 #include "NexNumber.h"
 
-NexNumber::NexNumber(uint8_t pid, uint8_t cid, const char *name)
-    :NexTouch(pid, cid, name)
+NexNumber::NexNumber(uint8_t pid, uint8_t cid, const char *name, const NexObject* page)
+    :NexTouch(pid, cid, name, page)
 {
 }
 
 bool NexNumber::getValue(uint32_t *number)
 {
     String cmd = String("get ");
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".val";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -34,7 +36,7 @@ bool NexNumber::setValue(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".val=";
     cmd += buf;
 
@@ -46,7 +48,7 @@ uint32_t NexNumber::Get_background_color_bco(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".bco";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -58,14 +60,14 @@ bool NexNumber::Set_background_color_bco(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".bco=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd="";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -74,7 +76,7 @@ uint32_t NexNumber::Get_font_color_pco(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".pco";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -86,14 +88,14 @@ bool NexNumber::Set_font_color_pco(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".pco=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -102,7 +104,7 @@ uint32_t NexNumber::Get_place_xcen(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".xcen";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -114,14 +116,14 @@ bool NexNumber::Set_place_xcen(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".xcen=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -130,7 +132,7 @@ uint32_t NexNumber::Get_place_ycen(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".ycen";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -142,14 +144,14 @@ bool NexNumber::Set_place_ycen(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".ycen=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -158,7 +160,7 @@ uint32_t NexNumber::getFont(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".font";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -170,14 +172,14 @@ bool NexNumber::setFont(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".font=";
     cmd += buf;
     sendCommand(cmd.c_str());
 
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -186,7 +188,7 @@ uint32_t NexNumber::Get_number_lenth(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".lenth";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -198,14 +200,14 @@ bool NexNumber::Set_number_lenth(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".lenth=";
     cmd += buf;
     sendCommand(cmd.c_str());
 
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -214,7 +216,7 @@ uint32_t NexNumber::Get_background_crop_picc(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".picc";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -226,14 +228,14 @@ bool NexNumber::Set_background_crop_picc(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".picc=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -241,7 +243,7 @@ bool NexNumber::Set_background_crop_picc(uint32_t number)
 uint32_t NexNumber::Get_background_image_pic(uint32_t *number)
 {
     String cmd = String("get ");
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".pic";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -253,14 +255,14 @@ bool NexNumber::Set_background_image_pic(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".pic=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }