+++ NexObject.cpp
@@ -5,6 +5,8 @@
  *
  * @author  Wu Pengfei (email:<pengfei.wu@itead.cc>)
  * @date    2015/8/13
+ * @author Jyrki Berg 2/17/2019 (https://github.com/jyberg)
+ * 
  * @copyright 
  * Copyright (C) 2014-2015 ITEAD Intelligent Systems Co., Ltd. \n
  * This program is free software; you can redistribute it and/or
@@ -14,11 +16,9 @@
  */
 #include "NexObject.h"
 
-NexObject::NexObject(uint8_t pid, uint8_t cid, const char *name)
+NexObject::NexObject(uint8_t pid, uint8_t cid, const char *name, const NexObject* page):
+__pid{pid},__cid{cid},__name{name}, __page{page}
 {
-    this->__pid = pid;
-    this->__cid = cid;
-    this->__name = name;
 }
 
 uint8_t NexObject::getObjPid(void)
@@ -31,11 +31,29 @@ uint8_t NexObject::getObjCid(void)
     return __cid;
 }
 
-const char* NexObject::getObjName(void)
+const char* NexObject::getObjName(void) const
 {
     return __name;
 }
 
+const char* NexObject::getObjPageName(void)
+{
+    if(__page)
+    {
+        return __page->getObjName();
+    }
+    return nullptr;
+}
+
+void NexObject::getObjGlobalPageName(String &gName)
+{
+    if(__page)
+    {
+        gName += __page->getObjName();
+        gName += ".";
+    }
+    gName +=__name;
+}
 void NexObject::printObjInfo(void)
 {
     dbSerialPrint("[");
@@ -45,6 +63,14 @@ void NexObject::printObjInfo(void)
     dbSerialPrint(",");
     dbSerialPrint(__cid);
     dbSerialPrint(",");
+    if(__page)
+    {
+        dbSerialPrint(__page->getObjName());
+    }
+    else
+    {
+        dbSerialPrint("(null)");
+    }    
     if (__name)
     {
         dbSerialPrint(__name);