+++ NexSlider.cpp
@@ -5,6 +5,8 @@
  *
  * @author  Wu Pengfei (email:<pengfei.wu@itead.cc>)
  * @date    2015/8/13
+ * @author Jyrki Berg 2/17/2019 (https://github.com/jyberg)
+ * 
  * @copyright 
  * Copyright (C) 2014-2015 ITEAD Intelligent Systems Co., Ltd. \n
  * This program is free software; you can redistribute it and/or
@@ -14,15 +16,15 @@
  */
 #include "NexSlider.h"
 
-NexSlider::NexSlider(uint8_t pid, uint8_t cid, const char *name)
-    :NexTouch(pid, cid, name)
+NexSlider::NexSlider(uint8_t pid, uint8_t cid, const char *name, const NexObject* page)
+    :NexTouch(pid, cid, name, page)
 {
 }
 
 bool NexSlider::getValue(uint32_t *number)
 {
     String cmd = String("get ");
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".val";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -34,7 +36,7 @@ bool NexSlider::setValue(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".val=";
     cmd += buf;
 
@@ -46,7 +48,7 @@ uint32_t NexSlider::Get_background_color_bco(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".bco";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -58,14 +60,14 @@ bool NexSlider::Set_background_color_bco(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".bco=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd="";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -74,7 +76,7 @@ uint32_t NexSlider::Get_font_color_pco(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".pco";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -86,14 +88,14 @@ bool NexSlider::Set_font_color_pco(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".pco=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -102,7 +104,7 @@ uint32_t NexSlider::Get_pointer_thickness_wid(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".wid";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -114,14 +116,14 @@ bool NexSlider::Set_pointer_thickness_wid(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".wid=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -130,7 +132,7 @@ uint32_t NexSlider::Get_cursor_height_hig(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".hig";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -142,14 +144,14 @@ bool NexSlider::Set_cursor_height_hig(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".hig=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -158,7 +160,7 @@ uint32_t NexSlider::getMaxval(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".maxval";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -170,14 +172,14 @@ bool NexSlider::setMaxval(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".maxval=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
@@ -186,7 +188,7 @@ uint32_t NexSlider::getMinval(uint32_t *number)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".minval";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
@@ -198,14 +200,14 @@ bool NexSlider::setMinval(uint32_t number)
     String cmd;
     
     utoa(number, buf, 10);
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".minval=";
     cmd += buf;
     sendCommand(cmd.c_str());
 	
     cmd = "";
     cmd += "ref ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }