+++ NexVariable.cpp
@@ -4,7 +4,9 @@
  * The implementation of class NexText. 
  *
  * @author  huang xiaoming (email:<xiaoming.huang@itead.cc>)
- * @date    2016/9/13
+ * @date    2016/9/
+ * @author Jyrki Berg 2/17/2019 (https://github.com/jyberg)
+ * 
  * @copyright 
  * Copyright (C) 2014-2015 ITEAD Intelligent Systems Co., Ltd. \n
  * This program is free software; you can redistribute it and/or
@@ -14,39 +16,48 @@
  */
 #include "NexVariable.h"
 
-NexVariable::NexVariable(uint8_t pid, uint8_t cid, const char *name)
-    :NexTouch(pid, cid, name)
+NexVariable::NexVariable(uint8_t pid, uint8_t cid, const char *name, const NexObject* page)
+    :NexTouch(pid, cid, name, page)
 {
 }
 
-uint32_t NexVariable::getValue(uint32_t *number)
+uint32_t NexVariable::getValue(int32_t *number)
 {
     String cmd = String("get ");
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".val";
     sendCommand(cmd.c_str());
     return recvRetNumber(number);
 }
 
-bool NexVariable::setValue(uint32_t number)
+bool NexVariable::setValue(int32_t number)
 {
     char buf[10] = {0};
     String cmd;
     
-    utoa(number, buf, 10);
-    cmd += getObjName();
+    itoa(number, buf, 10);
+    getObjGlobalPageName(cmd);
     cmd += ".val=";
     cmd += buf;
 
     sendCommand(cmd.c_str());
     return recvRetCommandFinished();
 }
-
-uint32_t NexVariable::getText(char *buffer, uint32_t len)
+bool NexVariable::getText(String &str)
 {
     String cmd;
     cmd += "get ";
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
+    cmd += ".txt";
+    sendCommand(cmd.c_str());
+    return recvRetString(str);
+}
+
+bool NexVariable::getText(char *buffer, uint16_t &len)
+{
+    String cmd;
+    cmd += "get ";
+    getObjGlobalPageName(cmd);
     cmd += ".txt";
     sendCommand(cmd.c_str());
     return recvRetString(buffer,len);
@@ -55,7 +66,7 @@ uint32_t NexVariable::getText(char *buffer, uint32_t len)
 bool NexVariable::setText(const char *buffer)
 {
     String cmd;
-    cmd += getObjName();
+    getObjGlobalPageName(cmd);
     cmd += ".txt=\"";
     cmd += buffer;
     cmd += "\"";