+++ release_notes.md
@@ -1,6 +1,20 @@
 # Release Notes
 
 --------------------------------------------------------------------------------
+# Release v0.10.0
+
+  - version: v0.10.0
+  - base: v0.9.0
+## Brief
+Enhanced and corrected Nextion library version for Arduino/Esp8266/NodeMcu, based on original Nextion Arduino library
+
+- Added support for NodeMcu/esp8266, Software serial, Software serial can be used with arduino also
+- Added support for global Nextion objects. (Optional page parameter added in the components)
+- NexVariable corrected to use int32_t data type
+- NextText corrected to return tru/false, and string length is returned in len parameter
+- NextText String object support added
+- other small bug fixes done
+- Added to support global Nextion events like CurrentPageIdCallback,systemStartUpCallback,...
 
 # Release v0.9.0